# doc/restapi/restapi.swagger2.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1",
    "title": "API V1"
  },
  "basePath": "/",
  "paths": {
...
    "/api/Game/{id}": {
      "delete": {
        "tags": [ "Game" ],
        "summary": "Delete Game with the ID provided.\r\n            \r\n            Example Usage: DELETE api/game/1",
        "operationId": "ApiGameByIdDelete",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Game ID.",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
...
  },
...
  "securityDefinitions": {}
}